chore: moves database initialisation to main
Allow tests to import components without firebase initialisation.
